+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## (IN PROGRESS)
 
+* Org list - Move focus to results pane when response has come. Refs UIORGS-208
 * Accessibility Error: Buttons must have discernible text. Refs-UIORGS-210.
 
 ## [2.2.0](https://github.com/folio-org/ui-organizations/tree/v2.2.0) (2020-10-09)