+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## (IN PROGRESS)
 
+## [2.2.1](https://github.com/folio-org/ui-organizations/tree/v2.2.1) (2020-11-10)
+[Full Changelog](https://github.com/folio-org/ui-organizations/compare/v2.2.0...v2.2.1)
+
 * Org list - Move focus to results pane when response has come. Refs UIORGS-208
 * Accessibility Error: Buttons must have discernible text. Refs-UIORGS-210.